Moving A2Q quantizer
