+++ src/brevitas/quant/scaled_int.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 from brevitas.core.function_wrapper import TensorClamp
-from brevitas.core.scaling.standalone import ParameterFromRuntimeStatsScaling
 from brevitas.quant.base import *
 from brevitas.quant.solver.act import ActQuantSolver
 from brevitas.quant.solver.bias import BiasQuantSolver
@@ -33,7 +32,9 @@ __all__ = [
     'Uint8ActPerTensorFloatBatchQuant1d',
     'Int8ActPerTensorFloatBatchQuant1d',
     'Uint8ActPerTensorFloatBatchQuant2d',
-    'Int8ActPerTensorFloatBatchQuant2d']
+    'Int8ActPerTensorFloatBatchQuant2d',
+    'Int8AccumulatorAwareWeightQuant',
+    'Int8WeightNormL2PerChannelFixedPoint']
 
 
 class Int8ActPerTensorFloatMinMaxInit(IntQuant,
@@ -398,3 +399,37 @@ class Int8ActPerTensorFloatBatchQuant1d(IntQuant,
         >>> act = QuantIdentity(act_quant=Int8ActPerTensorFloatBatchQuant1d)
     """
     pass
+
+
+class Int8WeightNormL2PerChannelFixedPoint(WeightNormPerChannelFloatDecoupled):
+    """
+    Experimental 8-bit narrow signed integer quantizer with learned per-channel scaling factors
+    and L2 weight normalization based on `Quantized Neural Networks for Low-Precision Accumulation
+    with Guaranteed Overflow Avoidance` by I. Colbert, A. Pappalardo, and J. Petri-Koenig
+    (https://arxiv.org/abs/2301.13376). The quantizer learns scaling factors in the float domain and
+    learns vector parameter g in the log domain with the half-way rounding function. Suitable for
+    retraining from floating-point depthwise separable weights.
+
+    Examples:
+        >>> from brevitas.nn import QuantConv2d
+        >>> conv = QuantConv2d(4, 4, 3, groups=4, weight_quant=Int8WeightNormL2PerChannelFixedPoint)
+        >>> conv.quant_weight()
+    """
+    bit_width = 8
+
+
+class Int8AccumulatorAwareWeightQuant(AccumulatorAwareWeightQuant):
+    """
+    Experimental 8-bit narrow signed accumulator-aware integer quantizer with learned per-channel
+    scaling factors based on `Quantized Neural Networks for Low-Precision Accumulation with Guaranteed
+    Overflow Avoidance` by I.Colbert, A.Pappalardo, and J.Petri-Koenig (https://arxiv.org/abs/2301.13376).
+    The quantizer learns scaling factors in the float domain and learns vector parameter g in the log
+    domain with the round-to-zero rounding function. The norm is clamped according the the specified
+    accumulator bit-width. Suitable for retraining from floating-point depthwise separable weights.
+
+    Examples:
+        >>> from brevitas.nn import QuantConv2d
+        >>> conv = QuantConv2d(4, 4, 3, groups=4, weight_quant=Int8AccumulatorAwareWeightQuant)
+        >>> conv.quant_weight()
+    """
+    bit_width = 8
